fibonacci-882-platform/backend/requirements.txt
# backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
websockets==12.0
httpx==0.25.2
pandas==2.1.4
numpy==1.26.2
pydantic==2.5.2
python-dotenv==1.0.0
aiofiles==23.2.1
apscheduler==3.10.4
aiohttp==3.9.1
aiohttp==3.9.1
